@@ -16,12 +16,12 @@
 
 package it.codingjam.github.ui.repo
 
-import android.arch.core.executor.testing.InstantTaskExecutorRule
-import android.support.annotation.StringRes
-import android.support.test.InstrumentationRegistry
-import android.support.test.espresso.Espresso.onView
-import android.support.test.espresso.assertion.ViewAssertions.matches
-import android.support.test.espresso.matcher.ViewMatchers.*
+import androidx.annotation.StringRes
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import androidx.test.InstrumentationRegistry
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.matcher.ViewMatchers.*
 import com.nhaarman.mockitokotlin2.mock
 import it.codingjam.github.ViewLibModule
 import it.codingjam.github.core.RepoDetail
