@@ -1,8 +1,6 @@
 package it.codingjam.github
 
-import android.support.design.widget.Snackbar
-import android.support.v4.app.Fragment
-import android.support.v4.app.FragmentActivity
+import com.google.android.material.snackbar.Snackbar
 import it.codingjam.github.core.RepoId
 import it.codingjam.github.ui.common.navigate
 import it.codingjam.github.ui.repo.RepoFragment
@@ -13,15 +11,15 @@ import javax.inject.Singleton
 @Singleton
 class AndroidNavigationController @Inject constructor() : NavigationController {
 
-    override fun navigateToRepo(fragment: Fragment, repoId: RepoId) {
+    override fun navigateToRepo(fragment: androidx.fragment.app.Fragment, repoId: RepoId) {
         RepoFragment.navigate(fragment, repoId)
     }
 
-    override fun navigateToUser(fragment: Fragment, login: String) {
+    override fun navigateToUser(fragment: androidx.fragment.app.Fragment, login: String) {
         UserFragment.navigate(fragment, login)
     }
 
-    override fun showError(activity: FragmentActivity, error: String?) {
+    override fun showError(activity: androidx.fragment.app.FragmentActivity, error: String?) {
         Snackbar.make(activity.findViewById(android.R.id.content), error
                 ?: "Error", Snackbar.LENGTH_LONG).show()
     }
