@@ -16,15 +16,15 @@
 
 package it.codingjam.github.ui.repo
 
-import android.arch.core.executor.testing.InstantTaskExecutorRule
 import android.content.SharedPreferences
 import android.os.Bundle
-import android.support.test.espresso.Espresso.onView
-import android.support.test.espresso.action.ViewActions.pressKey
-import android.support.test.espresso.action.ViewActions.typeText
-import android.support.test.espresso.assertion.ViewAssertions.matches
-import android.support.test.espresso.matcher.ViewMatchers.*
 import android.view.KeyEvent
+import androidx.arch.core.executor.testing.InstantTaskExecutorRule
+import androidx.test.espresso.Espresso.onView
+import androidx.test.espresso.action.ViewActions.pressKey
+import androidx.test.espresso.action.ViewActions.typeText
+import androidx.test.espresso.assertion.ViewAssertions.matches
+import androidx.test.espresso.matcher.ViewMatchers.*
 import com.nalulabs.prefs.fake.FakeSharedPreferences
 import com.nhaarman.mockitokotlin2.mock
 import it.codingjam.github.R
