@@ -16,14 +16,13 @@
 
 package it.codingjam.github.ui.common
 
-import android.databinding.ViewDataBinding
-import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import androidx.databinding.ViewDataBinding
 
 abstract class DataBoundViewHolder<T : Any, out V : ViewDataBinding>
 private constructor(val binding: V) :
-        RecyclerView.ViewHolder(binding.root) {
+        androidx.recyclerview.widget.RecyclerView.ViewHolder(binding.root) {
 
     constructor(parent: ViewGroup, factory: (LayoutInflater, ViewGroup, Boolean) -> V) :
             this(factory(LayoutInflater.from(parent.context), parent, false))
